Refactor to use a single JSON config file (#13)
In a followup, we want to demonstrate Protobuf. There, we will define a Protobuf for a Config.

This PR creates a similar `Config` object through pure Python. It also replaces reading .txt files with reading a single JSON file. 

In the process, we generalize `lang.py` and `greet.py` to no longer read hardcoded files. They are now parametrized, which makes their code simpler and more conventional Python. All reading of resources is consolidated.