Remove Python 2 (#56)
Demonstrating Python 2 is no longer pulling its weight. Python 2 interpreters are becoming harder to come by, and there are some issues with it like PyYAML failing to be built on macOS Big Sur when using a Python 2 interpreter.

Instead, we might want a dedicated example repo for Python 2 -> 3 repos. It can be more focused, such as not showing off Protobuf.