Upgrade to 2.2 and improve explanation of targets (#61)
Thanks to 2.2, we can now use the simplified design for `pex_binary` and `python_awslambda`. They no longer have `sources` fields, but have fields that act like it.

We can also simplify thanks to dep inference for Protobuf.

In the process, we do a better job explaining what each target is precisely doing. This is easier to do now that we have cleared up the responsibility of each target type.