Upgrade to Pants 2.5.0rc2. (#71)
And update constraints and python versions.

Confirmed that the obvious things work:

`./pants test ::`
`./pants lint ::`

etc.

`./pants binary ::` fails on MacOS because it cannot resolve
for Linux, as required by the Lambda target, and the resolve attempt
takes a looooong time. So we probably want to fail-fast in this case.
But otherwise, things seem stable.