Remove Protobuf and AWS Lambda from example (#84)
We got feedback in Slack that these two backends were making the example too complex. 

> I could not use that repo as a template because of the aws and protobuf stuff. ie do you want the repo to be a reusable template? or a fully baked example?

We would rather have the example repo be very easy to get started than to show every bell and whistle of Pants. Our docs are for the niche features like AWS Lambdas. Our example repo is to allow you to very easily try out Pants to get a feel for it before committing to prototyping in your own repo.