Simplify example to not use `translate` requirement (#85)
As explained in https://github.com/pantsbuild/example-python/pull/84, we think it's valuable for this example repository to be simple. In particular, we are demoing Pants itself, not the underlying code. 

This simplifies the project to use our own hardcoded `translate.json`, rather than the much more robust `translate` requirement. We still demonstrate the same things `translate` was showing for us:

* third-party management
* loading resource files

There's an added benefit that the repository is now much faster to build, especially on Apple Silicon where `lxml` must be compiled from source. Likewise, I no longer have to force Pants to use Py39+ for this to build on my M1! The default of 3.7+ now works.