Add MacOS/Apple Silicon guidance for changes to pants.toml (#98)
It looks me a couple of attempts to successfully follow along from the README.md on a MacBook Pro with Apple Silicon., ultimately coming down to needing to properly configure pants.toml to use pyenv and update the Python interpreter dependency. I believe the statement added to the README.md by this commit will help others in the same situation.